requirements.txt
streamlit
torch
torchvision
scikit-learn
pillow
matplotlib
seaborn
opencv-python
numpy